Add Logf
I have been using this library for a few week now, and I'm mostly happy, but I really miss a Logf.

And having it implmented here is much more effective than making the caller doing the `foo.Log(logg.Strings(fmt.Sprint...` dance
all the time.

The string building is performed by `fmt.Sprintf()`, but that will not be called if the level is not met.